Move preprocessor to project package
