+++ sit-cv-core/src/main/java/io/sitoolkit/cv/core/domain/project/Project.java
@@ -7,7 +7,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import io.sitoolkit.cv.core.domain.classdef.javaparser.preprocess.PreProcessor;
+import io.sitoolkit.cv.core.domain.project.preprocess.PreProcessor;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.Setter;
